refactor(cli): split runtime lifecycle tests
